requirements.txt
torch
torchvision
numpy
pillow
pytest
ruff
onnx
onnxscript
onnxruntime
onnx
onnxruntime
